AC-770: Separated total cost calculation
